Added twest cases for checking sequence of quads result in the right entities being passed to the index.
This is a weak but useful test that catches accidental uses
of DELETE for ADD and vice-versa.

Changed ExtendedEntity to have a BatchState.Mode mode field
recording whether it is generated by an ADD or DELETE
(or a NONE for cases where it doesn't matter).

Also gave it an equals() method.